Add type annotations to Python code

+++ python/argo/connection.py
@@ -1,19 +1,27 @@
+"""Utilities for connecting to Argo-based servers."""
+
 import json
+import os
 import re
 import socket
 import subprocess
-from . import netstring
-import sys
+import sys # pylint: disable=unused-import
+from typing import Any, Dict, List, Mapping, Optional, Union
+from mypy_extensions import TypedDict
 
-__doc__ = """Utilities for connecting to Argo-based servers."""
+from . import netstring
+
 
 # Must be boxed separately to enable sharing of connections
-class IDSource:
+class IDSource: # pylint: disable=too-few-public-methods
     """A source of unique identifiers."""
-    def __init__(self):
+
+    next_id : int
+
+    def __init__(self) -> None:
         self.next_id = 0
 
-    def get(self):
+    def get(self) -> int:
         """Get a unique (for this instance) number."""
         self.next_id += 1
         return self.next_id
@@ -22,7 +30,9 @@ class ServerProcess:
     """A wrapper around a server process, responsible for setting it up and
        killing it when finished."""
 
-    def __init__(self, command):
+    proc : Optional[subprocess.Popen]
+
+    def __init__(self, command : str) -> None:
         """Start the process using the given command.
 
            :param command: The command to be executed, using a shell, to start the server."""
@@ -30,7 +40,8 @@ class ServerProcess:
         self.proc = None
         self.setup()
 
-    def get_environment(self):
+    def get_environment(self) -> Optional[Union[Mapping[bytes, Union[bytes, str]],
+                                                Mapping[str, Union[bytes, str]]]]:
         """Return the environment in which the server process should be
            started. By default, this is the Override this method to
            allow customization. If ``None`` is returned, then the
@@ -38,7 +49,7 @@ class ServerProcess:
         """
         return None
 
-    def setup(self):
+    def setup(self) -> None:
         """Start a process, if one is not already running, using the
            environment determined by ``get_environment``."""
         if self.proc is None or self.proc.poll() is not None:
@@ -50,17 +61,19 @@ class ServerProcess:
                 # stderr=sys.stdout,
                 stderr=subprocess.DEVNULL,
                 env=self.get_environment(),
-                text=True)
+                universal_newlines=True)
             out_line = self.proc.stdout.readline()
 
             match = re.match(r'PORT (\d+)', out_line)
             if match:
                 self.port = int(match.group(1))
             else:
-                raise Exception("Failed to load process, output was `" + out_line + "' but expected PORT then a port.")
+                raise Exception("Failed to load process, output was `" + \
+                                out_line + "' but expected PORT then a port.")
 
-    def __del__(self):
-        self.proc.kill()
+    def __del__(self) -> None:
+        if self.proc is not None:
+            self.proc.kill()
 
 
 class ServerConnection:
@@ -74,7 +87,9 @@ class ServerConnection:
        answer received (if any). Furthermore, it has a means of
        sending messages to the server.
     """
-    def __init__(self, process):
+    replies : Dict[int, Any]
+
+    def __init__(self, process : ServerProcess) -> None:
         """:param process: The ``ServerProcess`` used for the connection."""
         self.process = process
         self.port = self.process.port
@@ -86,12 +101,12 @@ class ServerConnection:
         self.replies = {}
         self.ids = IDSource()
 
-    def get_id(self):
+    def get_id(self) -> int:
         """Return a fresh request ID that has not previously been used with
            this connection."""
         return self.ids.get()
 
-    def _buffer_replies(self):
+    def _buffer_replies(self) -> None:
         """Read any replies that the server has sent, and add their byte
            representation to the internal buffer, freeing up space in
            the pipe or socket.
@@ -105,28 +120,28 @@ class ServerConnection:
         except BlockingIOError:
             return None
 
-    def _get_one_reply(self):
+    def _get_one_reply(self) -> Optional[str]:
         """If a complete reply has been buffered, parse it from the buffer and
            return it as a bytestring."""
         try:
             (msg, rest) = netstring.decode(self.buf)
-            self.buf = rest
+            self.buf = bytearray(rest)
             return msg
         except (ValueError, IndexError):
             return None
 
-    def _process_replies(self):
+    def _process_replies(self) -> None:
         """Remove all pending replies from the internal buffer, parse them
            into JSON, and add them to the internal collection of replies.
         """
         self._buffer_replies()
-        r = self._get_one_reply()
-        while r is not None:
-            the_reply = json.loads(r)
+        reply_bytes = self._get_one_reply()
+        while reply_bytes is not None:
+            the_reply = json.loads(reply_bytes)
             self.replies[the_reply['id']] = the_reply
-            r = self._get_one_reply()
+            reply_bytes = self._get_one_reply()
 
-    def send_message(self, method, params):
+    def send_message(self, method : str, params : dict) -> int:
         """Send a message to the server with the given JSONRPC method and
            parameters. The return value is the unique request ID that
            was used for the message, which can be used to find
@@ -142,7 +157,7 @@ class ServerConnection:
         self.sock.send(msg_bytes)
         return request_id
 
-    def wait_for_reply_to(self, request_id):
+    def wait_for_reply_to(self, request_id : int) -> Any:
         """Block until a reply is received for the given
            ``request_id``. Return the reply."""
         self._process_replies()
@@ -153,4 +168,3 @@ class ServerConnection:
             finally:
                 self.sock.setblocking(False)
         return self.replies[request_id]
-
